requirements.txt
django1.10
python3.4.3
